+++ src/lib/error/ErrorKinds.ts
@@ -4,7 +4,6 @@ export const ErrorKinds = {
   ...ErrorKindsBase,
   DEPRECATION_CHECK_ERROR: 'DeprecationCheckError',
   EDITION_REQUEST_ERROR: 'EditionRequestError',
-  EVOLUTION_MANAGER_REPORT_ERROR: 'EvolutionManagerReportError',
   INVALID_OR_EXPIRED_TOKEN_ERROR: 'InvalidOrExpiredTokenError',
   OUTDATED_CHECK_ERROR: 'OutdatedCheckError',
   SETUP_TOOLING_ERROR: 'SetupToolingError',
@@ -13,4 +12,6 @@ export const ErrorKinds = {
   SSE_ERROR: 'SSEError',
   TELEMETRY_REPORTER_ERROR: 'TelemetryReporterError',
   TOOLBELT_CONFIG_MESSAGES_ERROR: 'ToolbeltConfigMessagesError',
+  APP_LOGS_SSE_ERROR: 'LogsSSEError',
+  APP_LOGS_PARSE_ERROR: 'LogsParseError',
 }